Refactor DOM plugin system to single module (#18025)
