--- packages/react-dom/src/events/DOMEventPluginOrder.js
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-/**
- * Module that is injectable into `EventPluginHub`, that specifies a
- * deterministic ordering of `EventPlugin`s. A convenient way to reason about
- * plugins, without having to package every one of them. This is better than
- * having plugins be ordered in the same order that they are injected because
- * that ordering would be influenced by the packaging order.
- * `ResponderEventPlugin` must occur before `SimpleEventPlugin` so that
- * preventing default on events is convenient in `SimpleEventPlugin` handlers.
- */
-const DOMEventPluginOrder = [
-  'ResponderEventPlugin',
-  'SimpleEventPlugin',
-  'EnterLeaveEventPlugin',
-  'ChangeEventPlugin',
-  'SelectEventPlugin',
-  'BeforeInputEventPlugin',
-];
-
-export default DOMEventPluginOrder;
