+++ packages/react-native-renderer/src/ReactNativeBridgeEventPlugin.js
@@ -28,9 +28,6 @@ const {
 const ReactNativeBridgeEventPlugin = {
   eventTypes: {},
 
-  /**
-   * @see {EventPluginHub.extractEvents}
-   */
   extractEvents: function(
     topLevelType: TopLevelType,
     targetInst: null | Object,
