+++ scripts/error-codes/codes.json
@@ -98,8 +98,8 @@
   "96": "EventPluginRegistry: Cannot inject event plugins that do not exist in the plugin ordering, `%s`.",
   "97": "EventPluginRegistry: Event plugins must implement an `extractEvents` method, but `%s` does not.",
   "98": "EventPluginRegistry: Failed to publish event `%s` for plugin `%s`.",
-  "99": "EventPluginHub: More than one plugin attempted to publish the same event name, `%s`.",
-  "100": "EventPluginHub: More than one plugin attempted to publish the same registration name, `%s`.",
+  "99": "EventPluginRegistry: More than one plugin attempted to publish the same event name, `%s`.",
+  "100": "EventPluginRegistry: More than one plugin attempted to publish the same registration name, `%s`.",
   "101": "EventPluginRegistry: Cannot inject event plugin ordering more than once. You are likely trying to load more than one copy of React.",
   "102": "EventPluginRegistry: Cannot inject two different event plugins using the same name, `%s`.",
   "103": "executeDirectDispatch(...): Invalid `event`.",
